requirements.txt
openai==0.27.0
fastapi
uvicorn
psycopg2
gradio
python-dotenv